Moved template spec to .h

+++ src/ejercicio5.cpp
@@ -18,7 +18,7 @@
 #include "user_io.h"
 
 template<std::size_t SIZE>
-std::optional<int> PositionOfFirst7(const std::array<int, SIZE> arr) {
+std::optional<size_t> PositionOfFirst7(const std::array<int, SIZE> arr) {
   for (size_t i{0}; i < SIZE; ++i)
     if (arr[i] == 7) return i;
   return {};
